Additions and fixes
Additions:
-StringUtils bracedStr
-FsUtils::createDir (tested on Windows and Linux)
-Color::fromStringB16
-JSON::parseVec3f
-Array-like accessor for vec3
-Atlas can now load images from multiple paths
-Atlas images can now be a JSON file with a set of elements and colormods, all added together.
-Camera3d can now have dimensions separate from the parent SDL_Window
-Camera::setPerspective (third person POVs)
-Mesh frustum caching
-Input::isKeyPressed()
-SQL_Time, SQLIte, libcrypto.so, libssl.so provided

Fixes:
-getWithoutExtension() can't handle files without extension
-getManyDirContents duplicate entries
-Typo in Box3::contains
-Atlas not padded
-BufferObj, GeoUtils, Quad, Tri, Vertex not nch namespaced
-Camera looking straight-up causes incorrect rotations with objects in a 3D world
-VertexAttribPointer issue in Draw:tri/texturedTri
-Incorrect frustum calculation in Meshes w/ scale!=1
-No move constructor or copy assignment for Text
-MainLoopDriver dooes not have benchmark tick op like with benchmark draw op

Removed:
-NTexture

+++ include/nch/opengl-utils/z/Draw.cpp
@@ -37,10 +37,10 @@ void Draw::texturedTri(const Vec3f& p0, const Vec3f& p1, const Vec3f& p2, const 
     glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 8*sf, (void*)0);
     glEnableVertexAttribArray(0);
     //1st arg 1 = Color attribute
-    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 8*sf, (void*)(3*sf));
+    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 8*sf, (void*)(uintptr_t)(3*sf));
     glEnableVertexAttribArray(1);
     //1st arg 2 = Texture coord attribute
-    glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, 8*sf, (void*)(6*sf));
+    glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, 8*sf, (void*)(uintptr_t)(6*sf));
     glEnableVertexAttribArray(2);
 
 	//Draw
@@ -101,7 +101,7 @@ void Draw::tri(const Vec3f& p0, const Vec3f& p1, const Vec3f& p2, GLuint& glVAO,
     glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 6 * sf, (void*)0);
     glEnableVertexAttribArray(0);
     // Color attribute
-    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 6 * sf, (void*)(3 * sf));
+    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 6 * sf, (void*)(uintptr_t)(3 * sf));
     glEnableVertexAttribArray(1);
 
 	//Draw
